Struct rustdoc_features_example::a::A

pub struct A {
    pub a: u8,
    pub foo: u8,
    pub foobar: u8,
}

A public struct that is always included in the crate.

Fields
a: u8

A public field that is always included in the crate.

foo: u8

A public field that is only included in the crate with feature foo.

foobar: u8

A public field that is only included in the crate with features foo and bar.

Methods

impl A

Some impls (from a) that are always included in the crate.

pub fn a(&self)

A public method that is always included in the crate.

pub fn foo(&self)

A public method that is only included in the crate with feature foo.

pub fn foobar2(&self)

A public method that is only included in the crate with features foo and bar.

impl A

Some impls (from bar) that are always included in the crate. (if the parent module is)

pub fn bar(&self)

A public method that is always included in the crate. (if the parent module is)

pub fn foobar(&self)

A public method that is only included in the crate with feature foo. (if the parent module is)