Trait rustdoc_features_example::a::FooTrait

pub trait FooTrait {
    fn foo(&self) { ... }

    fn foobar(&self) { ... }
}

A public trait that is only included in the crate with feature foo.

Provided Methods

fn foo(&self)

A trait method that is always included in the crate. (if the parent trait is)

fn foobar(&self)

A trait method that is only included in the crate with feature bar. (if the parent trait is)