Module rustdoc_features_example::foo

A public module that is only included in the crate with feature foo.

Structs

Foo
A public struct that is always included in the crate. (if the parent module is)

Traits

FooBarTrait
A public trait that is only included in the crate with feature bar. (if the parent module is)

FooTrait
A public trait that is always included in the crate. (if the parent module is)

Functions

foo
A public function that is always included in the crate. (if the parent module is)

foobar
A public function that is only included in the crate with feature bar. (if the parent module is)