Struct rustdoc_features_example::foo::Foo

pub struct Foo {
    pub foo: u8,
    pub foobar: u8,
}

A public struct that is always included in the crate. (if the parent module is)

Fields
foo: u8

A public field that is always included in the crate. (if the parent module is)

foobar: u8

A public field that is only included in the crate with feature bar. (if the parent module is)

Methods

impl Foo

fn foo(&self)
foo

A public method that is always included in the crate. (if the parent module is)

fn foobar(&self)
foo, bar

A public method that is only included in the crate with feature bar. (if the parent module is)